pyproject.toml
[tool.poetry]
name = "timekeeper"
version = "0.1.0"
description = ""
authors = ["Sajo Mathews <sajo@sajomathews.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-fasthtml = "^0.3.1"
pytest = "^8.3.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
ruff = "^0.5.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
